backend/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
